# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Admission Route API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/me/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_me_profile_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Put Profile",
        "operationId": "put_profile_me_profile_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/achievements": {
      "post": {
        "tags": [
          "achievements"
        ],
        "summary": "Create Achievement",
        "operationId": "create_achievement_me_achievements_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AchievementIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/achievements/{achievement_id}": {
      "patch": {
        "tags": [
          "achievements"
        ],
        "summary": "Patch Achievement",
        "operationId": "patch_achievement_me_achievements__achievement_id__patch",
        "parameters": [
          {
            "name": "achievement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Achievement Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AchievementPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "achievements"
        ],
        "summary": "Delete Achievement",
        "operationId": "delete_achievement_me_achievements__achievement_id__delete",
        "parameters": [
          {
            "name": "achievement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Achievement Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/demo": {
      "post": {
        "tags": [
          "demo"
        ],
        "summary": "Load Demo",
        "operationId": "load_demo_me_demo_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/reset": {
      "post": {
        "tags": [
          "demo"
        ],
        "summary": "Reset",
        "operationId": "reset_me_reset_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ai/passport": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Passport",
        "operationId": "passport_ai_passport_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ai/explain": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Explain",
        "operationId": "explain_ai_explain_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ai/roadmap-text": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Roadmap Text",
        "operationId": "roadmap_text_ai_roadmap_text_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoadmapTextIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapTextOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AchievementIn": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "SAT",
              "IELTS",
              "TOEFL",
              "OLYMPIAD",
              "PROJECT",
              "VOLUNTEER",
              "COMPETITION",
              "OTHER"
            ],
            "title": "Type"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "school",
                  "city",
                  "national",
                  "international"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "status": {
            "type": "string",
            "enum": [
              "done",
              "planned"
            ],
            "title": "Status",
            "default": "done"
          }
        },
        "type": "object",
        "required": [
          "type",
          "date"
        ],
        "title": "AchievementIn"
      },
      "AchievementPatch": {
        "properties": {
          "type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "SAT",
                  "IELTS",
                  "TOEFL",
                  "OLYMPIAD",
                  "PROJECT",
                  "VOLUNTEER",
                  "COMPETITION",
                  "OTHER"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "school",
                  "city",
                  "national",
                  "international"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "done",
                  "planned"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "AchievementPatch"
      },
...
      "ComputeResponse": {
        "properties": {
          "result": {
            "$ref": "#/components/schemas/RecommendationResult"
          },
          "roadmap": {
            "$ref": "#/components/schemas/Roadmap"
          },
          "diff": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Diff"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "result",
          "roadmap",
          "diff"
        ],
        "title": "ComputeResponse"
      },
...
      "ExplainIn": {
        "properties": {
          "university_id": {
            "type": "string",
            "title": "University Id"
          }
        },
        "type": "object",
        "required": [
          "university_id"
        ],
        "title": "ExplainIn"
      },
      "ExplainOut": {
        "properties": {
          "university_id": {
            "type": "string",
            "title": "University Id"
          },
          "summary": {
            "type": "string",
            "maxLength": 500,
            "title": "Summary"
          },
          "generated": {
            "type": "boolean",
            "title": "Generated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "university_id",
          "summary"
        ],
        "title": "ExplainOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PassportOut": {
        "properties": {
          "goal": {
            "$ref": "#/components/schemas/PassportItem"
          },
          "strengths": {
            "items": {
              "$ref": "#/components/schemas/PassportItem"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Strengths"
          },
          "constraints": {
            "items": {
              "$ref": "#/components/schemas/PassportItem"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 1,
            "title": "Constraints"
          },
          "risk": {
            "$ref": "#/components/schemas/PassportItem"
          },
          "generated": {
            "type": "boolean",
            "title": "Generated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "goal",
          "strengths",
          "constraints",
          "risk"
        ],
        "title": "PassportOut"
      },
...
      "Priorities": {
        "properties": {
          "cost": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Cost",
            "default": 0.5
          },
          "prestige": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Prestige",
            "default": 0.5
          },
          "location": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Location",
            "default": 0.5
          },
          "aid": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Aid",
            "default": 0.5
          }
        },
        "type": "object",
        "title": "Priorities"
      },
      "Profile": {
        "properties": {
          "grade": {
            "type": "integer",
            "enum": [
              10,
              11,
              12
            ],
            "title": "Grade"
          },
          "gpa5": {
            "type": "number",
            "maximum": 5.0,
            "minimum": 2.0,
            "title": "Gpa5"
          },
          "majors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Majors"
          },
          "countries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Countries"
          },
          "budget_per_year_usd": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Budget Per Year Usd"
          },
          "needs_aid": {
            "type": "boolean",
            "title": "Needs Aid"
          },
          "intake_year": {
            "type": "integer",
            "title": "Intake Year"
          },
          "priorities": {
            "$ref": "#/components/schemas/Priorities"
          },
          "achievements": {
            "items": {
              "$ref": "#/components/schemas/Achievement"
            },
            "type": "array",
            "title": "Achievements",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "grade",
          "gpa5",
          "majors",
          "countries",
          "budget_per_year_usd",
          "needs_aid",
          "intake_year",
          "priorities",
          "created_at"
        ],
        "title": "Profile"
      },
      "ProfileIn": {
        "properties": {
          "grade": {
            "type": "integer",
            "enum": [
              10,
              11,
              12
            ],
            "title": "Grade"
          },
          "gpa5": {
            "type": "number",
            "maximum": 5.0,
            "minimum": 2.0,
            "title": "Gpa5"
          },
          "majors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Majors"
          },
          "countries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Countries"
          },
          "budget_per_year_usd": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Budget Per Year Usd"
          },
          "needs_aid": {
            "type": "boolean",
            "title": "Needs Aid"
          },
          "intake_year": {
            "type": "integer",
            "title": "Intake Year"
          },
          "priorities": {
            "$ref": "#/components/schemas/Priorities"
          },
          "initial_achievements": {
            "items": {
              "$ref": "#/components/schemas/AchievementIn"
            },
            "type": "array",
            "title": "Initial Achievements",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "grade",
          "gpa5",
          "majors",
          "countries",
          "budget_per_year_usd",
          "needs_aid",
          "intake_year",
          "priorities"
        ],
        "title": "ProfileIn",
        "description": "Profile body for PUT /me/profile. Achievements are managed separately, but\nonboarding may send initial ones (without ids)."
      },
...
      "RoadmapTextIn": {
        "properties": {
          "step_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 40,
            "title": "Step Ids"
          }
        },
        "type": "object",
        "required": [
          "step_ids"
        ],
        "title": "RoadmapTextIn"
      },
      "RoadmapTextOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StepText"
            },
            "type": "array",
            "title": "Items"
          },
          "generated": {
            "type": "boolean",
            "title": "Generated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "RoadmapTextOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}